// src/webhook/enums/webhook.enum.ts
export enum WebhookType {
  INITIATED = 'request-initiated',
  BID = 'validate-blockchain-id',
  TOKEN = 'validate-token',
  ON_CHAIN_WALLET = 'validate-on-chain-wallet',
  ORDERID = 'validate-order-id',
  HASH = 'validate-hash',
  AMOUNT = 'validate-minimum-amount',
  CREATE = 'validate-create-order',
  DEPOSIT_SETTINGS = 'validate-deposite-settings',
  WALLET = 'validate-wallet',
  WALLET_TRANSACTION = 'validate-wallet-transaction',
  PLATFORM = 'validate-platform-id',
  DEPOSIT_AND_STAKE_SETTINGS = 'validate-deposite-and-stake-settings',
  DEPOSIT_AND_STAKE_NO_MACHINE = 'no-available-machine-for-stake',
  STAKE_NOT_ENABLED = 'stake-not-enabled',
  NODE_K_SETTINGS_NOT_FOUND = 'node-settings-not-found',
  MISMATCH_TOKEN = 'mismatch-in-stake-token-and-machine-token',
  EXCEEDED_STAKE_LIMIT = 'exceeded-stake-limit',
  UNEXPECTED_ERROR = 'unexpected-error',
  COMPLETED = 'completed',
  UNEXPECTED = 'validate-unexpected-error',
  USER_DATA_UPDATE = 'user-data-update',
  MINIMUM_LIMIT_NOT_MET = 'exceeded-minumum-limit',
  ADDRESS_EXISTS = 'address_exists',
  NETWORK = 'network-not-found',
  COIN_TO_USDT = 'converting-coin-to-usd',
  TOKEN_MISMATCH = 'token-mismatch',
  MEMBER_SHIP_EXPIRE_DATE = 'membership_expire_date',
  MEMBER_SHIP_STATUS = 'membership_status',
  MEMBER_SHIP_UNEXPECTED_ERROR = 'membership_unexpected-error',
  MACHINE_HIGHTEST_COLLATORAL_ERROR = 'validate-machine-hightest-collatoral',
}

export enum WebhookMessages {
  INITIATED = 'webhook initiated successfully',
  BID = 'Invalid Blockchain ID',
  TOKEN = 'Invalid Token',
  ORDERID = 'Order id is already exists.',
  HASH = 'The request with the same hash already exists.',
  COMPLETED = 'Webhook processed successfully',
  CREATE = 'Something went wrong. Your order could not be created.',
  AMOUNT = 'The amount is less than the minimum required for deposit',
  UNEXPECTED = 'An unexpected error occurred',
  WALLET = 'No wallet found for this user and token',
  WALLET_TRANSACTION = 'Failed to create wallet transaction',
  DEPOSIT_SETTINGS = 'No settings are available for the requested token.',
  PLATFORM = 'Invalid platform.',
  ON_CHAIN_WALLET = 'No user found with this wallet address.',
  DEPOSIT_AND_STAKE_SETTINGS = 'No deposit and stake settings are available for the requested token.',
  DEPOSIT_AND_STAKE_NO_MACHINE = `User don't have such machine to stake.`,
  STAKE_NOT_ENABLED = 'Staking is currently disabled.',
  NODE_K_SETTINGS_NOT_FOUND = 'Node k settings are not found.',
  MISMATCH_TOKEN = 'Mismatch occured in deposit token and stake token in machine.',
  EXCEEDED_STAKE_LIMIT = 'Exceeded the stake limit.',
  UNEXPECTED_ERROR = 'Unexpected error occured while trying to stake.',
  MINIMUM_LIMIT_NOT_MET = 'Amount requested is less than the minimum required.',
  ADDRESS_EXISTS = 'Same address is not allowed for multiple deposits.',
  DEPOSIT_REQUEST_CANCELLED_BY_USER = 'Request successfully cancelled by the user',
  NO_DEPOSIT_RECEIVED = 'No deposit received within the active period',
  NETWORK = 'Network is not available ',
  COIN_TO_USDT = 'Failed to convert coin to USD',
  TOKEN_MISMATCH = 'Token mismatch detected',
  MEMBER_SHIP_STATEUS = 'The user membership status update was not completed',
  MEMBER_SHIP_STATEUS_COMPLETE = 'The user membership status updated',
  MEMBER_SHIP_UNEXPECTED_ERROR = 'Unexpected error occured while trying to Updagre Membership.',
}

export enum WebhookPlatform {
  JB_HOMNIFI = 'jb',
}

export enum WebhookDataStatus {
  SUCCESS = 'success',
  FAILED = 'failed',
}